# Use blobs instead of random integers 