--- a/napari/benchmarks/benchmark_labels_layer.py
+++ b/napari/benchmarks/benchmark_labels_layer.py
@@ -9,7 +9,9 @@ import numpy as np
 from napari.components.dims import Dims
 from napari.layers import Labels
 
-from .utils import Skiper
+from .utils import Skiper, gen_blobs
+
+MAX_VAL = 2**23
 
 
 class Labels2DSuite:
@@ -23,7 +25,8 @@ class Labels2DSuite:
 
     def setup(self, n, dtype):
         np.random.seed(0)
-        self.data = np.random.randint(20, size=(n, n), dtype=dtype)
+        self.data = gen_blobs((n, n), dtype, blob_count=int(np.log2(n) ** 2))
+        # np.random.randint(20, size=(n, n), dtype=dtype)
         self.layer = Labels(self.data)
         self.layer._raw_to_displayed(self.data, (slice(0, n), slice(0, n)))
 
@@ -84,7 +87,9 @@ class LabelsDrawing2DSuite:
 
     def setup(self, n, brush_size, color_mode, contour):
         np.random.seed(0)
-        self.data = np.random.randint(64, size=(n, n), dtype=np.int32)
+        self.data = gen_blobs(
+            (n, n), np.int32, blob_count=int(np.log2(n) ** 2)
+        )
 
         colors = None
         if color_mode == 'direct':
@@ -117,12 +122,7 @@ class Labels2DColorDirectSuite(Labels2DSuite):
             raise NotImplementedError("Skip on PR (speedup)")
         np.random.seed(0)
         info = np.iinfo(dtype)
-        self.data = np.random.randint(
-            low=max(-10000, info.min),
-            high=min(10000, info.max),
-            size=(n, n),
-            dtype=dtype,
-        )
+        self.data = gen_blobs((n, n), dtype, blob_count=int(np.log2(n) ** 2))
         random_label_ids = np.random.randint(
             low=max(-10000, info.min), high=min(10000, info.max), size=20
         )
@@ -148,7 +148,9 @@ class Labels3DSuite:
             raise NotImplementedError("Skip on CI (not enough memory)")
 
         np.random.seed(0)
-        self.data = np.random.randint(20, size=(n, n, n), dtype=dtype)
+        self.data = gen_blobs(
+            (n, n, n), dtype, blob_count=int(np.log2(n) ** 2)
+        )
         self.layer = Labels(self.data)
         self.layer._slice_dims(Dims(ndim=3, ndisplay=3))
         self.layer._raw_to_displayed(