--- a/napari/benchmarks/benchmark_labels_layer.py
+++ b/napari/benchmarks/benchmark_labels_layer.py
@@ -9,7 +9,7 @@ import numpy as np
 from napari.components.dims import Dims
 from napari.layers import Labels
 
-from .utils import Skiper, gen_blobs
+from .utils import Skiper, labeled_particles
 
 MAX_VAL = 2**23
 
@@ -25,7 +25,9 @@ class Labels2DSuite:
 
     def setup(self, n, dtype):
         np.random.seed(0)
-        self.data = gen_blobs((n, n), dtype, blob_count=int(np.log2(n) ** 2))
+        self.data = labeled_particles(
+            (n, n), dtype=dtype, n=int(np.log2(n) ** 2), seed=1
+        )
         # np.random.randint(20, size=(n, n), dtype=dtype)
         self.layer = Labels(self.data)
         self.layer._raw_to_displayed(self.data, (slice(0, n), slice(0, n)))
@@ -87,8 +89,8 @@ class LabelsDrawing2DSuite:
 
     def setup(self, n, brush_size, color_mode, contour):
         np.random.seed(0)
-        self.data = gen_blobs(
-            (n, n), np.int32, blob_count=int(np.log2(n) ** 2)
+        self.data = labeled_particles(
+            (n, n), dtype=np.int32, n=int(np.log2(n) ** 2), seed=1
         )
 
         colors = None
@@ -122,7 +124,9 @@ class Labels2DColorDirectSuite(Labels2DSuite):
             raise NotImplementedError("Skip on PR (speedup)")
         np.random.seed(0)
         info = np.iinfo(dtype)
-        self.data = gen_blobs((n, n), dtype, blob_count=int(np.log2(n) ** 2))
+        self.data = labeled_particles(
+            (n, n), dtype=dtype, n=int(np.log2(n) ** 2), seed=1
+        )
         random_label_ids = np.random.randint(
             low=max(-10000, info.min), high=min(10000, info.max), size=20
         )
@@ -148,8 +152,8 @@ class Labels3DSuite:
             raise NotImplementedError("Skip on CI (not enough memory)")
 
         np.random.seed(0)
-        self.data = gen_blobs(
-            (n, n, n), dtype, blob_count=int(np.log2(n) ** 2)
+        self.data = labeled_particles(
+            (n, n, n), dtype=dtype, n=int(np.log2(n) ** 2), seed=1
         )
         self.layer = Labels(self.data)
         self.layer._slice_dims(Dims(ndim=3, ndisplay=3))