--- a/napari_builtins/_ndims_balls.py
+++ b/napari_builtins/_ndims_balls.py
@@ -1,9 +1,12 @@
+import numpy as np
+
 from napari.benchmarks.utils import labeled_particles
 
 
 def labeled_particles2d():
+    seed = np.random.default_rng().integers(np.iinfo(np.int64).max)
     labels, density, points = labeled_particles(
-        (1024, 1024), return_density=True
+        (1024, 1024), seed=seed, return_density=True
     )
 
     return [
@@ -14,12 +17,13 @@ def labeled_particles2d():
 
 
 def labeled_particles3d():
+    seed = np.random.default_rng().integers(np.iinfo(np.int64).max)
     labels, density, points = labeled_particles(
-        (256, 512, 512), return_density=True
+        (256, 512, 512), seed=seed, return_density=True
     )
 
     return [
-        (density, {"name": "density"}, "image"),
-        (labels, {"name": "labels"}, "labels"),
-        (points, {"name": "points"}, "points"),
+        (density, {"name": "density", "metadata": {"seed": seed}}, "image"),
+        (labels, {"name": "labels", "metadata": {"seed": seed}}, "labels"),
+        (points, {"name": "points", "metadata": {"seed": seed}}, "points"),
     ]