--- a/napari_builtins/builtins.yaml
+++ b/napari_builtins/builtins.yaml
@@ -120,6 +120,12 @@ contributions:
     - id: napari.data.text
       title: Generate text sample
       python_name: napari_builtins._skimage_data:text
+    - id: napari.data.balls_2d
+      title: Generate 2d_balls sample
+      python_name: napari_builtins._ndims_balls:labeled_particles2d
+    - id: napari.data.balls_3d
+      title: Generate 3d_balls sample
+      python_name: napari_builtins._ndims_balls:labeled_particles3d
 
   readers:
     - command: napari.get_reader
@@ -291,3 +297,9 @@ contributions:
     - display_name: Text
       key: text
       command: napari.data.text
+    - display_name: 2D Balls
+      key: balls_2d
+      command: napari.data.balls_2d
+    - display_name: 3D Balls
+      key: balls_3d
+      command: napari.data.balls_3d