# Description

During investigation of https://github.com/scikit-image/scikit-image/issues/7262 I have created a blobs function that is quite fast and produces data more realistic than just `np.random.randint`. 

So I suggest adopting it in benchmarks. 

It is slightly faster on my machine 

main:
```
[50.00%] ·· ======================================================== ================
                                   benchmark                          total duration
            -------------------------------------------------------- ----------------
             benchmark_labels_layer.LabelsDrawing2DSuite.time_draw        33.9s
                 benchmark_labels_layer.Labels3DSuite.mem_layer           16.7s
             benchmark_labels_layer.Labels3DSuite.time_create_layer       16.5s
               benchmark_labels_layer.Labels3DSuite.time_refresh          16.5s
                 benchmark_labels_layer.Labels3DSuite.mem_data            16.5s
                                      ...                                  ...
                                     total                                6.92m
            ======================================================== ================           
```

This branch: 
```
[50.00%] ·· ============================================================ ================
                                     benchmark                            total duration
            ------------------------------------------------------------ ----------------
               benchmark_labels_layer.LabelsDrawing2DSuite.time_draw          33.6s
                   benchmark_labels_layer.Labels2DSuite.time_fill             20.4s
             benchmark_labels_layer.Labels2DColorDirectSuite.time_fill        18.1s
               benchmark_labels_layer.Labels3DSuite.time_create_layer         13.5s
             benchmark_labels_layer.Labels3DSuite.time_raw_to_displayed       13.4s
                                        ...                                    ...
                                       total                                  6.56m
            ============================================================ ================
```

It also adds Balls 2D and Balls 3D example data with importing from `napari.examples.utils`