@Czaki 2023-12-26
> Is it worth moving to a more general location in the code base?

yes and no

It may be a useful example, but without cherry-picking this to 0.4.19, we need to use only utils.py content in benchmark to be able to compare main against the latest release.