@psobolewskiPhD 2023-12-26
Ok. Maybe we make it a sample or something in 0.5.0. Can make a task issue once merged.