@Czaki 2023-12-26
> Ok. Maybe we make it a sample or something in 0.5.0. Can make a task issue once merged.

We could make it sample now. But it will require to have two copies of code. Or we agree on cherry pick it to 0.4.19.