@psobolewskiPhD 2023-12-28
The samples are amazing, especially with points and labels!
😍 
I think that following the existing Samples, these should be after astronaught as Balls and Balls (3D)

BTW The 3D one has some issue with background in mip though:
<img width="1217" alt="image" src="https://github.com/napari/napari/assets/76622105/2249b6db-3de6-42bb-8c9d-eb2dc1605f7e">

Some locations arn't set a value perhaps? or are those just the only 0 values...