@Czaki 2023-12-28
> BTW The 3D one has some issue with background in mip though:

It is not an issue. The background is visible only for 0 value. 
I have improved it a little.