@Czaki 2024-01-02
I have added tests