@Czaki 2024-01-02
And found that the code was not python 3.8 compatible.