@jni 2024-01-03
@Czaki this is for 0.5.0 so 3.8 compatibility is not important?