@jni 2024-01-03
I see that the CI still uses 3.8, I guess because we wanted to keep it until 0.4.19. I'm a little shocked that the walrus operator there is reported as a syntax error, since it was introduced in 3.8? Is it because it's not allowed in square brackets due to implementation details?